LayoutTests/fast/events/monotonic-event-time-real-gap-expected.txt

This test checks that DOM event timestamps advance to reflect real wall-clock time between events, even without an explicit eventSender.leapForward() call.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS firstTimeStamp is defined.
PASS secondTimeStamp is defined.
PASS (secondTimeStamp - firstTimeStamp) > 50 is true
PASS (secondTimeStamp - firstTimeStamp) < 5000 is true
PASS successfullyParsed is true

TEST COMPLETE
